requirements.txt
# Core dependencies
python-dotenv>=0.19.0
pydantic>=1.8.2

# AI and ML
openai>=1.0.0
langchain>=0.0.300

# Deep Learning and NLP
torch>=1.9.0
transformers>=4.11.0
sentence-transformers>=2.2.0

# Data processing
numpy>=1.21.0
pandas>=1.5.0
scikit-learn>=1.0.2

# NLP
spacy>=3.5.0
nltk>=3.8.1

# Utilities
typing-extensions>=3.10.0
python-dateutil>=2.8.2
tqdm>=4.65.0

# Code Analysis and Generation
black>=22.3.0
astor>=0.8.1
joblib>=1.1.0

# Vector Stores and Embeddings
faiss-cpu>=1.7.4
openai-embeddings>=0.1.0

# Monitoring and Performance
prometheus-client>=0.16.0
psutil>=5.9.0

# Visualization and Analytics
plotly>=5.13.0
dash>=2.9.0
networkx>=2.8.0

# Testing and Development
pytest>=7.0.0
pytest-asyncio>=0.18.0
pytest-cov>=3.0.0
mypy>=0.950

# Advanced Analytics
prophet>=1.1.0
